Notifications/FormattableContent: Code organization
