+++ WordPress/Classes/ViewRelated/Notifications/FormattableContent/FooterContentGroup.swift
@@ -1,0 +1,4 @@
+
+class FooterContentGroup: FormattableContentGroup {
+
+}
